requirements/base.txt
# Django
Django==4.2.7
djangorestframework==3.14.0
python-decouple==3.8

# Database
psycopg2-binary==2.9.9

# WSGI Server
gunicorn==21.2.0
whitenoise==6.6.0

# CORS
django-cors-headers==4.3.1
